requirements.txt
bandit
black
flake8
flask
instructor
invoke
isort
mypy
openai
pydantic
python-dotenv
litellm
uuid6
shortuuid